Warning bofore deleting record added, styles for button added, colors improvement

+++ PartyListMain.py
@@ -44,7 +44,7 @@ class PartyWindow(QMainWindow):
         self.table_party_list.setModel(self.personModel.model)
         self.table_party_list.setSelectionBehavior(QAbstractItemView.SelectRows) 
         self.table_party_list.horizontalHeader().setStyleSheet('QHeaderView::section {background-color: rgb(69, 86, 206); color: rgb(250, 250, 250); font-weight: bold;} QToolTip { background-color: #8ad4ff; color: black; border: #8ad4ff solid 1px}')
-        self.table_party_list.verticalHeader().setStyleSheet('QHeaderView::section {background-color: rgb(69, 86, 206); color: rgb(250, 250, 250); font-weight: bold;} QToolTip { background-color: #8ad4ff; color: black; border: #8ad4ff solid 1px}')
+        self.table_party_list.verticalHeader().setStyleSheet('QHeaderView::section {background-color: rgb(60, 60, 60); color: rgb(250, 210, 210); font-size: 6} QToolTip { background-color: #8ad4ff; color: black; border: #8ad4ff solid 1px}')
 
         self.button_delete.clicked.connect(lambda: self.deletePerson())
         self.button_close.clicked.connect(app.closeAllWindows)
@@ -55,8 +55,11 @@ class PartyWindow(QMainWindow):
         row = self.table_party_list.currentIndex().row()  
         if row < 0:  
             return  
-        
-        self.personModel.deletePerson(row) 
+
+        messageBox = QMessageBox.warning(self, "Warning!", "<FONT COLOR='white'>Really do you want to delete the selected person?", QMessageBox.Ok | QMessageBox.Cancel)  
+  
+        if messageBox == QMessageBox.Ok:
+            self.personModel.deletePerson(row) 
 
 app = QApplication(sys.argv)
 connection = Database.createConnection("party_list.db")
@@ -65,4 +68,5 @@ widget = QtWidgets.QStackedWidget()
 widget.setWindowTitle('PARTY LIST')
 widget.addWidget(mainwindow)
 widget.show()
+app.setStyleSheet(Path('styles.qss').read_text())
 app.exec()
